chore: Move platform ifdefs to .cpp files and remove `#error`.
This allows us to always compile all .cpp files, making the build system
simpler while not increasing complexity in the source, i.e. this is a net
reduction of complexity. I can see some weak arguments for wanting strict
checking, but I don't see what bugs could be prevented by that
strictness.

This PR does not itself simplify the CMake build system, but opens the
door to doing so.